lingdb/requirements.txt
Django==2.1
django-appconf==1.0.2
django-crispy-forms==1.7.2
django-filter==2.0.0
django-floppyforms==1.7.0
django-select2==6.3.1
django-watson==1.5.2
psycopg2==2.7.5
psycopg2-binary==2.7.5
python-dateutil==2.8.0
pytz==2018.5
six==1.12.0
